Move ApplicationsInProgress to its own file
