+++ src/applications/personalization/dashboard-2/components/apply-for-benefits/ApplyForBenefits.jsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useMemo } from 'react';
-import moment from 'moment';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 
 import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
@@ -11,13 +10,6 @@ import {
   selectProfile,
 } from '~/platform/user/selectors';
 
-import {
-  formLinks,
-  formTitles,
-  isSIPEnabledForm,
-  presentableFormIDs,
-  sipFormSorter,
-} from '~/applications/personalization/dashboard/helpers';
 import { getEnrollmentStatus as getEnrollmentStatusAction } from '~/applications/hca/actions';
 import { HCA_ENROLLMENT_STATUSES } from '~/applications/hca/constants';
 import {
@@ -31,56 +23,8 @@ import {
   eduDirectDepositIsSetUp,
 } from '@@profile/selectors';
 
-import ApplicationInProgress from './ApplicationInProgress';
+import ApplicationsInProgress from './ApplicationsInProgress';
 import BenefitOfInterest from './BenefitOfInterest';
-
-const ApplicationsInProgress = ({ savedForms }) => {
-  const verifiedSavedForms = useMemo(
-    () => savedForms.filter(isSIPEnabledForm).sort(sipFormSorter),
-    [savedForms],
-  );
-
-  return (
-    <>
-      <h3 className="vads-u-font-size--h4 vads-u-font-family--sans vads-u-margin-bottom--2p5">
-        Applications in progress
-      </h3>
-      {verifiedSavedForms.length > 0 && (
-        <div className="vads-l-grid-container vads-u-padding--0">
-          <div className="vads-l-row">
-            {verifiedSavedForms.map(form => {
-              const formId = form.form;
-              const formTitle = `Application for ${formTitles[formId]}`;
-              const presentableFormId = presentableFormIDs[formId];
-              const { lastUpdated, expiresAt } = form.metadata || {};
-              const lastOpenedDate = moment
-                .unix(lastUpdated)
-                .format('MMMM D, YYYY');
-              const expirationDate = moment
-                .unix(expiresAt)
-                .format('MMMM D, YYYY');
-              const continueUrl = `${formLinks[formId]}resume`;
-              return (
-                <ApplicationInProgress
-                  key={formId}
-                  continueUrl={continueUrl}
-                  expirationDate={expirationDate}
-                  formId={formId}
-                  formTitle={formTitle}
-                  lastOpenedDate={lastOpenedDate}
-                  presentableFormId={presentableFormId}
-                />
-              );
-            })}
-          </div>
-        </div>
-      )}
-      {!verifiedSavedForms.length && (
-        <p>You have no applications in progress.</p>
-      )}
-    </>
-  );
-};
 
 const BenefitsOfInterest = ({ children, showChildren }) => {
   return (
